Update to match API changes

+++ src/main/java/net/como/client/misc/maths/Vec3.java
@@ -127,7 +127,7 @@ public class Vec3 {
     }
 
     public BlockPos toBlockPos() {
-        return new BlockPos(x, y, z);
+        return new BlockPos((int)x, (int)y, (int)z);
     }
 
     @Override
